fix: scrollbar css
